disp: msm: dsi: increase maximum number of allowed DSI modes to 64
Maximum number of DSI modes the driver currently supports is 32 but
some of the customers require more than 32 modes, hence increase to 64.

+++ msm/dsi/dsi_defs.h
@@ -655,7 +655,7 @@ struct dsi_display_mode_priv_info {
 	struct msm_ratio pclk_scale;
 	struct msm_roi_caps roi_caps;
 	bool widebus_support;
-	u32 allowed_mode_switch;
+	u64 allowed_mode_switch;
 	bool disable_rsc_solver;
 };
